+++ Ktisis/Overlay/Skeleton.cs
@@ -4,11 +4,11 @@ using System.Numerics;
 using ImGuiNET;
 using ImGuizmoNET;
 
-using Ktisis.Interface.Windows.Toolbar;
-using Ktisis.Interface.Windows.Workspace;
 using Ktisis.Structs;
 using Ktisis.Structs.Actor;
 using Ktisis.Structs.Bones;
+using Ktisis.Interface.Windows.Toolbar;
+using Ktisis.Interface.Windows.Workspace.Tabs;
 
 namespace Ktisis.Overlay {
 	public static class Skeleton {
@@ -65,7 +65,7 @@ namespace Ktisis.Overlay {
 
 			// Draw skeleton
 
-			var isUsing = ImGuizmo.IsUsing() || Workspace.Transform.IsActive || TransformWindow.Transform.IsActive;
+			var isUsing = ImGuizmo.IsUsing() || PoseTab.Transform.IsActive || TransformWindow.Transform.IsActive;
 
 			for (var p = 0; p < model->Skeleton->PartialSkeletonCount; p++) {
 				var partial = model->Skeleton->PartialSkeletons[p];
